Start moving all functionality that works directly on MIR into a separate module.
