+++ prusti-viper/src/encoder/mir/mod.rs
@@ -1,0 +1,4 @@
+//! Part of the encoder that converts from MIR level into `vir-high` that does
+//! not depend anymore on the compiler internals.
+
+pub(crate) mod pure_functions;
