+++ prusti-viper/src/encoder/mod.rs
@@ -21,7 +21,6 @@ mod mir_interpreter;
 mod name_interner;
 mod places;
 mod procedure_encoder;
-mod pure_functions;
 mod spec_encoder;
 mod spec_function_encoder;
 pub use spec_function_encoder::SpecFunctionKind;
@@ -31,6 +30,7 @@ mod type_encoder;
 mod utils;
 mod snapshot;
 mod mirror_function_encoder;
+mod mir;
 mod purifier;
 mod array_encoder;
 pub mod counterexample;
